fixes from pr #2 review
