+++ app/src/main/kotlin/com/streafy/pizzashift2024/pizzalist/data/model/PizzaModel.kt
@@ -8,9 +8,9 @@ data class PizzaModel(
     val id: String,
     val name: String,
     val ingredients: List<Ingredient>,
-    val toppings: List<Topping>,
+    val toppings: List<ToppingModel>,
     val description: String,
-    val sizes: List<Size>,
+    val sizes: List<SizeModel>,
     val doughs: List<Dough>,
     val calories: Int,
     val protein: String,
