+++ app/src/main/kotlin/com/streafy/pizzashift2024/pizzalist/data/model/SizeModel.kt
@@ -3,8 +3,7 @@ package com.streafy.pizzashift2024.pizzalist.data.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Topping(
-    val cost: Int,
-    val img: String,
-    val name: String
+data class SizeModel(
+    val name: String,
+    val price: Int
 )
