+++ app/src/main/kotlin/com/streafy/pizzashift2024/pizzalist/domain/Pizza.kt
@@ -1,7 +1,7 @@
 package com.streafy.pizzashift2024.pizzalist.domain
 
 data class Pizza(
-    val id: Int,
+    val id: String,
     val name: String,
     val description: String,
     val price: Int,
